modalのdisablePortalを無効化
